[Silver II] Title: 나무 자르기, Time: 224 ms, Memory: 4900 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.c
@@ -20,7 +20,7 @@ int main() {
 		for (int i = 0; i < N; i++) {
 			if (trees[i] <= m) continue;
 
-			len += (long long)(trees[i] - m);
+			len += trees[i] - m;
 		}
 
 		if (len < M) {
